# perf: avoid syncing directories when they already existed