When doing a Put, the containing directory is always created. If it already exist,
no error bubbles up from `makeDirNoSync` and later this directly is always synced on disk
in `makeDir`.

This generate FS operations which are almost always unnecessary as the supporting folders
are quickly all created.