@MichaelMure 2022-05-09
I tried to benchmark both version but the results vary widely from one run to another, even when using flags like `-benchtime 100x`.