@guseggert 2022-05-13
Are you saying that there's no measurable difference in performance? If that's the case, I'd prefer to keep things consistent / make less assumptions, makes the code easier to maintain.