@MichaelMure 2022-05-13
I did another benchmarking on NFS:
```
Before:

BenchmarkPut-12    	      19	  90327492 ns/op	    2442 B/op	      39 allocs/op
BenchmarkPut-12    	      20	  60113199 ns/op	    2934 B/op	      39 allocs/op
BenchmarkPut-12    	      19	 144812107 ns/op	    3016 B/op	      40 allocs/op
BenchmarkPut-12    	      20	  60713097 ns/op	    2507 B/op	      39 allocs/op
BenchmarkPut-12    	      18	  57389155 ns/op	    3008 B/op	      39 allocs/op
BenchmarkPut-12    	      10	 135447145 ns/op	    3417 B/op	      42 allocs/op
BenchmarkPut-12    	      19	  72482237 ns/op	    2962 B/op	      39 allocs/op
BenchmarkPut-12    	      20	  86261195 ns/op	    2867 B/op	      39 allocs/op
AVG                                                    88443203.375		2894.125		39.5

After:

BenchmarkPut-12    	      22	  61761422 ns/op	    2660 B/op	      36 allocs/op
BenchmarkPut-12    	      20	  81016432 ns/op	    2343 B/op	      36 allocs/op
BenchmarkPut-12    	      21	  79589929 ns/op	    2761 B/op	      36 allocs/op
BenchmarkPut-12    	      10	 121222088 ns/op	    2496 B/op	      39 allocs/op
BenchmarkPut-12    	      24	  57701723 ns/op	    2670 B/op	      35 allocs/op
BenchmarkPut-12    	      21	  47750888 ns/op	    2346 B/op	      36 allocs/op
BenchmarkPut-12    	      25	  82605228 ns/op	    2326 B/op	      35 allocs/op
BenchmarkPut-12    	      19	  84322739 ns/op	    2815 B/op	      36 allocs/op
AVG                                                    76996306.125		2552.125		36.125
```

13% faster, 12% less allocation (bytes), 9% less allocation (count)